Refactoring, formatting rules
